refactor:  api files to api/ dir in explainer
