+++ backend/explainer/api/urls.py
@@ -1,14 +1,14 @@
 from django.urls import path, include
 from rest_framework.routers import DefaultRouter
 
-from explainer import views
+from .views import NetworkViewSet
 
 
 app_name = 'explainer'
 
 # Create a router and register our viewsets with it.
 router = DefaultRouter()
-router.register(r'network', views.NetworkViewSet, 'network')
+router.register(r'network', NetworkViewSet, 'network')
 
 urlpatterns = [
     path('', include(router.urls)),
